backend/requirements/base.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary>=2.9.10
python-dotenv==1.0.1
Pillow==10.4.0
black==26.3.0